+++ Cargo.toml
@@ -16,14 +16,11 @@ keywords = ["fcm", "firebase", "notification"]
 edition = "2018"
 
 [features]
-default = ["native-tls", "gauth"]
+default = ["native-tls"]
 
 native-tls = ["reqwest/native-tls"]
 rustls = ["reqwest/rustls-tls"]
 vendored-tls = ["reqwest/native-tls-vendored"]
-
-gauth = ["dep:gauth"]
-yup-oauth2 = ["dep:yup-oauth2"]
 
 [dependencies]
 serde = { version = "1", features = ["derive"] }
@@ -33,9 +30,7 @@ reqwest = { version = "0.11", features = ["json"], default-features = false }
 chrono = "0.4"
 thiserror = "1"
 dotenvy = "0.15"
-
-gauth = { version = "0.7", optional = true }
-yup-oauth2 = { version = "9", optional = true }
+yup-oauth2 = "9"
 
 [dev-dependencies]
 tokio = { version = "1.0", features = ["rt-multi-thread", "macros"] }
